+++ Source/DependencyInversion/ServiceScopeFactory.cs
@@ -34,6 +34,6 @@ public class ServiceScopeFactory : IServiceScopeFactory
         return new TenantServiceScope(
             rootScope,
             _lifetimeScope.BeginLifetimeScope(builder =>
-                builder.RegisterSource(new UnknownServiceOnTenantContainerRegistrationSource(rootScope.ServiceProvider))));
+                builder.RegisterSource(new UnknownServiceOnTenantContainerRegistrationSource(rootScope.ServiceProvider, _lifetimeScope.ComponentRegistry.Registrations, false))));
     }
 }