+++ Source/SDK/Builders/DolittleClientConfiguration.cs
@@ -15,53 +15,57 @@ namespace Dolittle.SDK.Builders;
 /// <summary>
 /// Represents the <see cref="IDolittleClient"/> configuration.
 /// </summary>
-public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClientConfiguration
+public class DolittleClientConfiguration : IConfigurationBuilder
 {
     /// <summary>
-    /// Gets or sets the <see cref="Version"/> of the Head.
+    /// Gets the <see cref="Version"/> of the Head.
     /// </summary>
     public Version Version { get; private set; } = Version.NotSet;
 
     /// <summary>
-    /// Gets or sets the Runtime host.
+    /// Gets the Runtime host.
     /// </summary>
     public string RuntimeHost { get; private set; } = "localhost";
 
     /// <summary>
-    /// Gets or sets theRuntime port.
+    /// Gets the Runtime port.
     /// </summary>
     public ushort RuntimePort { get; private set; } = 50053;
 
     /// <summary>
-    /// Gets or sets the ping-interval <see cref="TimeSpan"/>.
+    /// Gets the ping-interval <see cref="TimeSpan"/>.
     /// </summary>
     public TimeSpan PingInterval { get; private set; } = TimeSpan.FromSeconds(5);
 
     /// <summary>
-    /// Gets or sets the event serializer provider.
+    /// Gets the event serializer provider.
     /// </summary>
     public Func<JsonSerializerSettings> EventSerializerProvider { get; private set; } = () => new JsonSerializerSettings();
 
+    /// <summary>
+    /// Gets the OpenTelemetry settings provider.
+    /// </summary>
     public Func<OpenTelemetrySettings> OpenTelemetrySettingsProvider { get; private set; } = () => new OpenTelemetrySettings();
 
     /// <summary>
-    /// Gets or sets the<see cref="ILoggerFactory"/>.
+    /// Gets the <see cref="ILoggerFactory"/>.
     /// </summary>
-    public ILoggerFactory LoggerFactory { get; private set; } = Microsoft.Extensions.Logging.LoggerFactory.Create(_ =>
-    {
-        _.SetMinimumLevel(LogLevel.Information);
-        _.AddConsole();
-    });
+    public ILoggerFactory? LoggerFactory { get; private set; }
 
     /// <summary>
-    /// Gets or sets the<see cref="IServiceProvider"/>.
+    /// Gets the <see cref="IServiceProvider"/>.
     /// </summary>
-    public IServiceProvider ServiceProvider { get; private set; }
+    public IServiceProvider? ServiceProvider { get; private set; }
 
     /// <summary>
-    /// Gets or sets the<see cref="ConfigureTenantServices"/> callback.
+    /// Gets the <see cref="ConfigureTenantServices"/> callback.
     /// </summary>
     public ConfigureTenantServices? ConfigureTenantServices { get; private set; }
+    
+    /// <summary>
+    /// Gets the <see cref="CreateTenantServiceProvider"/> factory.
+    /// </summary>
+    public CreateTenantServiceProvider? TenantServiceProviderFactory { get; private set; }
 
     /// <summary>
     /// Configures the <see cref="DolittleClientConfiguration"/> with the configuration values from <see cref="Configurations.Dolittle"/>.
@@ -71,18 +75,15 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
     public static DolittleClientConfiguration FromConfiguration(Configurations.Dolittle config)
     {
         var result = new DolittleClientConfiguration();
-        if (config.Runtime != default)
+        
+        if (!string.IsNullOrEmpty(config.Runtime.Host))
         {
-            var runtime = config.Runtime;
-            if (!string.IsNullOrEmpty(runtime.Host))
-            {
-                result.RuntimeHost = config.Runtime.Host;
-            }
+            result.RuntimeHost = config.Runtime.Host!;
+        }
 
-            if (runtime.Port.HasValue)
-            {
-                result.RuntimePort = runtime.Port.Value;
-            }
+        if (config.Runtime.Port.HasValue)
+        {
+            result.RuntimePort = config.Runtime.Port.Value;
         }
 
         if (config.PingInterval.HasValue)
@@ -95,7 +96,7 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
             result.Version = new VersionConverter().FromString(config.HeadVersion);
         }
 
-        if (config.Otlp is { })
+        if (config.Otlp is not null)
         {
             result.OpenTelemetrySettingsProvider = () => new OpenTelemetrySettings
             {
@@ -106,25 +107,15 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
 
         return result;
     }
-
-    /// <summary>
-    /// Sets the <see cref="Version"/> of the Head.
-    /// </summary>
-    /// <param name="version">The <see cref="Version"/>.</param>
-    /// <returns></returns>
+    
+    /// <inheritdoc />
     public IConfigurationBuilder WithVersion(Version version)
     {
         Version = version;
         return this;
     }
-
-    /// <summary>
-    /// Connect to a specific host and port for the Dolittle runtime.
-    /// </summary>
-    /// <param name="host">The host name to connect to.</param>
-    /// <param name="port">The port to connect to.</param>
-    /// <returns>the client configuration builder for continuation.</returns>
-    /// <remarks>If not specified, host 'localhost' and port 50053 will be used.</remarks>
+    
+    /// <inheritdoc />
     public IConfigurationBuilder WithRuntimeOn(string host, ushort port)
     {
         RuntimeHost = host;
@@ -132,23 +123,14 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
         return this;
     }
 
-    /// <summary>
-    /// Sets the <see cref="ILoggerFactory"/> to use for creating instances of <see cref="ILogger"/> for the client configuration.
-    /// </summary>
-    /// <param name="factory">The given <see cref="ILoggerFactory"/>.</param>
-    /// <returns>the client configuration builder for continuation.</returns>
-    /// <remarks>If not used, a factory with 'Trace' level logging will be used.</remarks>
+    /// <inheritdoc />
     public IConfigurationBuilder WithLogging(ILoggerFactory factory)
     {
         LoggerFactory = factory;
         return this;
     }
 
-    /// <summary>
-    /// Sets a callback that configures the <see cref="JsonSerializerSettings"/> for serializing events.
-    /// </summary>
-    /// <param name="jsonSerializerSettingsBuilder"><see cref="Action{T}"/> that gets called with <see cref="JsonSerializerSettings"/> to modify settings.</param>
-    /// <returns>the client configuration builder for continuation.</returns>
+    /// <inheritdoc />
     public IConfigurationBuilder WithEventSerializerSettings(Action<JsonSerializerSettings> jsonSerializerSettingsBuilder)
     {
         EventSerializerProvider = () =>
@@ -160,11 +142,7 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
         return this;
     }
 
-    /// <summary>
-    /// Set a callback for overriding OpenTelemetry settings
-    /// </summary>
-    /// <param name="openTelemetrySettingsBuilder"></param>
-    /// <returns></returns>
+    /// <inheritdoc />
     public IConfigurationBuilder WithOpenTelemetrySettings(Action<OpenTelemetrySettings> openTelemetrySettingsBuilder)
     {
         OpenTelemetrySettingsProvider = () =>
@@ -177,36 +155,32 @@ public class DolittleClientConfiguration : IConfigurationBuilder, IDolittleClien
         return this;
     }
 
-    /// <summary>
-    /// Sets the ping interval for communicating with the microservice.
-    /// </summary>
-    /// <param name="interval">The ping interval.</param>
-    /// <returns>the client configuration builder for continuation.</returns>
+    /// <inheritdoc />
     public IConfigurationBuilder WithPingInterval(TimeSpan interval)
     {
         PingInterval = interval;
         return this;
     }
 
-    /// <summary>
-    /// Configures the root <see cref="IServiceProvider"/> for the <see cref="IDolittleClient"/>.
-    /// </summary>
-    /// <param name="serviceProvider">The <see cref="IServiceProvider"/>.</param>
-    /// <returns>The client for continuation.</returns>
+    /// <inheritdoc />
     public IConfigurationBuilder WithServiceProvider(IServiceProvider serviceProvider)
     {
         ServiceProvider = serviceProvider;
+        
         return this;
     }
 
-    /// <summary>
-    /// Configures a <see cref="ConfigureTenantServices"/> callback for configuring the tenant specific IoC containers.
-    /// </summary>
-    /// <param name="configureTenantServices">The <see cref="ConfigureTenantServices"/> callback.</param>
-    /// <returns>The client for continuation.</returns>
+    /// <inheritdoc />
     public IConfigurationBuilder WithTenantServices(ConfigureTenantServices configureTenantServices)
     {
         ConfigureTenantServices = configureTenantServices;
+        return this;
+    }
+
+    /// <inheritdoc />
+    public IConfigurationBuilder WithTenantServiceProviderFactory(CreateTenantServiceProvider factory)
+    {
+        TenantServiceProviderFactory = factory;
         return this;
     }
 }